move get actual compression to test common
