+++ hub/core/chunk_engine/tests/test_benchmark_chunk_engine.py
@@ -10,9 +10,8 @@ from hub.core.tensor import (
     read_samples_from_tensor,
     create_tensor,
 )
-from hub.core.tests.common import TENSOR_KEY
 from hub.core.typing import StorageProvider
-from hub.tests.common import get_random_array
+from hub.tests.common import TENSOR_KEY, get_random_array
 from hub.tests.common_benchmark import (
     parametrize_benchmark_chunk_sizes,
     parametrize_benchmark_dtypes,
