+++ hub/core/tests/common.py
@@ -1,3 +1,6 @@
+from io import BytesIO
+from hub.constants import UNCOMPRESSED
+from PIL import Image, UnidentifiedImageError
 from hub.core.meta.tensor_meta import TensorMeta
 from hub.core.index import Index
 import pytest
@@ -12,12 +15,9 @@ from hub.core.tensor import (
     append_tensor,
     create_tensor,
     extend_tensor,
-    tensor_exists,
     read_samples_from_tensor,
 )
 from hub.core.typing import StorageProvider
-from hub.tests.common import TENSOR_KEY
-from hub.util.keys import get_chunk_key
 
 STORAGE_FIXTURE_NAME = "storage"
 DATASET_FIXTURE_NAME = "ds"
